rename serde module to serde_impl module
needed because of issues with rustfmt
